feat: get order by id

+++ src/main/java/com/learn/photo/repository/OrderProductRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface OrderProductRepository extends CrudRepository<OrderProduct, Integer> {
-}
+public interface OrderProductRepository extends CrudRepository<OrderProduct, Long> {}
