+++ src/main/java/com/learn/photo/repository/OrderRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface OrderRepository extends CrudRepository<Order, Integer> {
-}
+public interface OrderRepository extends CrudRepository<Order, Long> {}
