+++ src/main/java/com/learn/photo/repository/ProductRepository.java
@@ -4,5 +4,4 @@ import com.learn.photo.model.Photo;
 import com.learn.photo.model.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Integer> {
-}
+public interface ProductRepository extends CrudRepository<Product, Long> {}
